Add CRUD with basic CLI interface
This commit introduces a memdb subpackage that contains the backend
logic to handle our data using a hash map of strings as both its key and
value. Basic helper functions are constructed to allow easier interface
with the map.

An example CLI (main.go) in the root level of this project demonstrates
how one could implement go-memdb/memdb.

To-do: Implement transactions