Use string pointer for GET requests
A string pointer allows us to determine if the key/value pair exists
within out databse. Instead of manipulating the output to be "NULL" for
entries that don't exist within the memdb subpackge, we defer that
handling to the person implementing/using this package instead. Thus,
with a null string pointer, the developer can differentiate between the
object not existing, or if someone actually wrote the value "NULL" into
the db.